refactor: Refactor pending requests feature package structure
Moved PendingRequests-related classes from 'employees' to 'requests' and 'requests/components' packages for better organization. Updated repository import paths and renamed AppModule to PendingRequestsModule to reflect the new structure.
