+++ app/app/src/main/java/com/lojasocial/app/ui/requests/PendingRequestsViewModel.kt
@@ -1,8 +1,8 @@
-package com.lojasocial.app.ui.employees
+package com.lojasocial.app.ui.requests
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.lojasocial.app.data.repository.PendingRequestsRepository
+import com.lojasocial.app.repository.PendingRequestsRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
