+++ app/app/src/main/java/com/lojasocial/app/ui/requests/components/StatusBadge.kt
@@ -1,6 +1,5 @@
-package com.lojasocial.app.ui.employees
+package com.lojasocial.app.ui.requests.components
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
